+++ experimental/packages/opentelemetry-instrumentation/src/types_internal.ts
@@ -14,8 +14,7 @@
  * limitations under the License.
  */
 
-import { TracerProvider } from '@opentelemetry/api';
-import { MeterProvider } from '@opentelemetry/api-metrics';
+import { TracerProvider, MeterProvider } from '@opentelemetry/api';
 import { InstrumentationBase } from './platform';
 import { Instrumentation } from './types';